Added events structure page
